feat: SmartRoutingAddress component
# Conflicts:
#	packages/react-kit/src/smart-routing/hooks/useSmartRoutingAddress.ts

# Conflicts:
#	packages/react-kit/src/smart-routing/hooks/useSmartRoutingAddress.test.tsx
